Add initial github action
The intention is to move away from travis and appveyor which have
become very slow.

See: #2356

+++ check.py
@@ -329,7 +329,7 @@ def run_gcc_tests():
         print('run...', output_file)
         actual = subprocess.check_output([os.path.abspath(output_file)]).decode('utf-8')
         os.remove(output_file)
-        if sys.platform == 'darwin':
+        if sys.platform == 'darwin' and os.path.exists(output_file + '.dSYM'):
             # Also removes debug directory produced on Mac OS
             shutil.rmtree(output_file + '.dSYM')
